requirements.txt
# Core App & Data Handling
streamlit
pandas
numpy
yfinance
plotly
scikit-learn

# --- STABLE AI/ML CONFIGURATION ---
# Pinning specific versions known to be compatible on Streamlit Cloud
tensorflow==2.15.0
transformers==4.38.2
torch==2.1.0
safetensors==0.4.2

# API Libraries
google-generativeai
alpha-vantage
feedparser
streamlit-autorefresh

# Optional
openai